src/main/java/com/example/creditunion/account/service/AccountService.java
package com.example.creditunion.account.service;

import com.example.creditunion.account.model.Account;
import com.example.creditunion.account.repository.AccountRepository;
import com.example.creditunion.account.exception.InternalServerException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import java.util.List;
import java.util.Optional;

@Service
public class AccountService {

    @Autowired
    private AccountRepository accountRepository;

    public List<Account> getAllAccounts() {
        return accountRepository.findAll();
    }

    public Optional<Account> getAccountById(Long id) {
        return accountRepository.findById(id);
    }

    public Account createAccount(Account account) {
        return accountRepository.save(account);
    }

    public Account updateAccount(Long id, Account accountDetails) {
        Account account = accountRepository.findById(id)
                .orElseThrow(() -> new InternalServerException("Account not found with id " + id));
        account.setFirstName(accountDetails.getFirstName());
        account.setLastName(accountDetails.getLastName());
        account.setAddress(accountDetails.getAddress());
        return accountRepository.save(account);
    }

    public void deleteAccount(Long id) {
        Account account = accountRepository.findById(id)
                .orElseThrow(() -> new InternalServerException("Account not found with id " + id));
        accountRepository.delete(account);
    }

    public List<Account> searchAccountsByName(String name) {
        return accountRepository.findByFirstNameContainingIgnoreCaseOrLastNameContainingIgnoreCase(name, name);
    }

}
